// serverGaragiste/src/models/Garagiste.ts
/** @format */

export type GaragisteType = {
  id?: number; // Optional: as it's auto-generated by PostgreSQL
  name: string;
  email: string;
  password: string; // Remember to handle password hashing appropriately
  createdAt?: Date; // Optional: as it's auto-generated or set by trigger
  updatedAt?: Date; // Optional: as it's auto-generated or set by trigger
  phone?: string;
  profileImage?: string;
  deplomeImage?: string;
};